Moved auth code to its own module
This will make it easier to refactor into a more complete
authentication + authorization middleware
